Extract unit conversion methods to a separate file
