+++ lib/src/circular_slider.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:sleek_circular_slider/src/slider_animations.dart';
+import 'package:sleek_circular_slider/src/unit_conversions.dart';
 import 'utils.dart';
 import 'appearance.dart';
 import 'slider_label.dart';
